Move existing logic to display_cutout
